+++ packages/api/src/tx/modules/ecocredit/v1/create_batch_amino.ts
@@ -5,7 +5,7 @@ import {
   BatchIssuance,
   OriginTx,
 } from '../../../../generated/regen/ecocredit/v1/types';
-import { AminoDate } from '../converter-utils';
+import { AminoDate } from '../../converter-utils';
 
 const msgCreateBatchAminoType = 'regen/MsgCreateBatch';