+++ packages/api/src/tx/msg.ts
@@ -10,7 +10,10 @@ import { Registry, EncodeObject, GeneratedType } from '@cosmjs/proto-signing';
 
 import { SigningConnectionOptions } from '../api';
 import { createStargateSigningClient } from './stargate-signing';
-import { createEcocreditAminoConverters } from './modules';
+import {
+  createDataAminoConverters,
+  createEcocreditAminoConverters,
+} from './modules';
 import { messageTypeRegistry } from '../generated/typeRegistry';
 
 export interface MessageClient {
@@ -26,6 +29,7 @@ export interface MessageClient {
 function createDefaultTypes(): AminoConverters {
   return {
     ...createEcocreditAminoConverters(),
+    ...createDataAminoConverters(),
   };
 }