+++ packages/api/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.1](https://github.com/regen-network/regen-js/compare/v0.7.0...v0.7.1) (2024-08-05)
+
+
+### Bug Fixes
+
+* use base64 string for hash in amino ([88737e5](https://github.com/regen-network/regen-js/commit/88737e56f612fb957e997e26ff852662768c2046))
+
+
+
+
+
+# [0.7.0](https://github.com/regen-network/regen-js/compare/v0.6.0...v0.7.0) (2024-08-05)
+
+
+### Features
+
+* add amino data MsgAttest converter ([739163b](https://github.com/regen-network/regen-js/commit/739163b341ff466d14fdd7df8025ffd7c0cadfe4))
+
+
+
+
+
+
 # [0.6.0](https://github.com/regen-network/regen-js/compare/v0.3.0...v0.6.0) (2023-03-27)